feat: add `LegendDecorator` to support the legend of the graph.

+++ lib/core/algorithm/decorator/force_motion_decorator.dart
@@ -9,7 +9,7 @@ import 'package:flutter_graph_view/flutter_graph_view.dart';
 ///
 /// 将顶点的力转换成运动速度并改变节点位置的装饰器。
 class ForceMotionDecorator extends GraphAlgorithm {
-  double unitTime;
+  int unitTime;
   ForceMotionDecorator({this.unitTime = 1, super.decorators});
 
   Vector2 getForce(Vertex v) {
@@ -35,7 +35,7 @@ class ForceMotionDecorator extends GraphAlgorithm {
           return;
         } else {
           // s = 1/2 * a * t^2
-          var s = a * 0.5 * unitTime;
+          var s = a * 0.5 * unitTime.toDouble();
           v.position += s;
         }
       }
